+++ layrz_sdk/entities/checkpoints/__init__.py
@@ -1,4 +1,4 @@
-""" Checkpoints entities """
+"""Checkpoints entities"""
+
 from .checkpoint import Checkpoint
-from .geofence import Geofence
 from .waypoint import Waypoint